Deletes jQuery and replaces all animations and DOM access by Angular way
- Regarding this comment:
He uses JQuery -> There is no point in using JQuery in an Angular project as the framework provides ways to manipulate the DOM (ViewChild) in a more elegant way. Apart from that, JQuery is a really intrusive way of manipulating the DOM. It also increases the bundle size of the app.
- Also adjusts no-search and no-results divs styles to absolute positions